refactoring first and fourth swap steps
