+++ src/front/shared/pages/Swap/UTXOSwap/SwapList/SwapList.tsx
@@ -8,15 +8,14 @@ import styles from './SwapList.scss'
 
 import DepositWindow from '../DepositWindow/DepositWindow'
 
-import UTXOFirstStep from './UTXOSteps/FirstStep'
-import UTXOSecondStep from './UTXOSteps/SecondStep'
-import UTXOThirdStep from './UTXOSteps/ThirdStep'
-import UTXOFourthStep from './UTXOSteps/FourthStep'
+import FirstStep from './SwapSteps/FirstStep'
+import FourthStep from './SwapSteps/FourthStep'
 
-import ABFirstStep from './ABSteps/FirstStep'
-import ABSecondStep from './ABSteps/SecondStep'
-import ABThirdStep from './ABSteps/ThirdStep'
-import ABFourthStep from './ABSteps/FourthStep'
+import UTXOSecondStep from './SwapSteps/UTXOSteps/SecondStep'
+import UTXOThirdStep from './SwapSteps/UTXOSteps/ThirdStep'
+
+import ABSecondStep from './SwapSteps/ABSteps/SecondStep'
+import ABThirdStep from './SwapSteps/ABSteps/ThirdStep'
 
 
 import MakerUtxoToAbtTexts from './SwapProgressTexts/MakerUtxoToAb'
@@ -34,41 +33,7 @@ export default class SwapList extends Component<any, any> {
   constructor(props) {
     super(props)
 
-    const {
-      swap: {
-        sellCurrency,
-        flow: {
-          stepNumbers,
-          isTakerMakerModel,
-        },
-        flow,
-      },
-      fields,
-    } = props
-
-    const isTaker = flow.isTaker()
-
-    this._fields = fields
-
-    const { currencyName } = fields
-
-    const first = stepNumbers.sign
-    const second = sellCurrency === currencyName ? stepNumbers[`submit-secret`] : stepNumbers[`wait-lock-utxo`]
-    const fourth = sellCurrency === currencyName ? stepNumbers[`lock-utxo`] : stepNumbers[`sync-balance`]
-    const fifth = sellCurrency === currencyName ? stepNumbers[`wait-lock-eth`] : stepNumbers[`lock-eth`]
-    const sixth = sellCurrency === currencyName ? stepNumbers[`withdraw-eth`] : stepNumbers[`wait-withdraw-eth`]
-    const seventh = sellCurrency === currencyName ? stepNumbers.finish : stepNumbers[`withdraw-utxo`]
-    const eighth = sellCurrency === currencyName ? stepNumbers.end : stepNumbers.finish
-
-    this.state = {
-      first,
-      second,
-      fourth,
-      fifth,
-      sixth,
-      seventh,
-      eighth,
-    }
+    this._fields = props.fields
 
     console.group('SwapList >%c constructor', 'color: green;')
     console.log('fields: ', this._fields)
@@ -77,8 +42,12 @@ export default class SwapList extends Component<any, any> {
 
   getStepName = () => {
     const {
+      swap: {
+        flow: {
+          stepNumbers
+        }
+      },
       flow: {
-        stepNumbers,
         step
       }
      } = this.props
@@ -90,6 +59,8 @@ export default class SwapList extends Component<any, any> {
         stepName = key
       }
     })
+
+    console.log('stepName', stepName)
 
     return stepName
   }
@@ -110,29 +81,26 @@ export default class SwapList extends Component<any, any> {
       currencyData,
       tokenItems
     } = this.props
-    const { first, second, fourth, fifth, sixth, seventh, eighth } = this.state
 
     const { currencyName } = fields
 
     const isUTXOSide = flowClass.isUTXOSide
 
-    const FirstStep = (isUTXOSide) ? UTXOFirstStep : ABFirstStep
     const SecondStep = (isUTXOSide) ? UTXOSecondStep : ABSecondStep
     const ThirdStep = (isUTXOSide) ? UTXOThirdStep : ABThirdStep
-    const FourthStep = (isUTXOSide) ? UTXOFourthStep : ABFourthStep
 
-    const UtxoToAbTexts = (flow.isTaker()) ? TakerUtxoToAbTexts : MakerUtxoToAbtTexts
-    const AbToUtxoTexts = (flow.isTaker()) ? TakerAbToUtxoTexts : MakerAbToUtxoTexts
+    const UtxoToAbTexts = (flow.isTaker) ? TakerUtxoToAbTexts : MakerUtxoToAbtTexts
+    const AbToUtxoTexts = (flow.isTaker) ? TakerAbToUtxoTexts : MakerAbToUtxoTexts
 
     const swapTexts = (
       <Fragment>
         {
           ['BtcLikeToEth', 'BtcLikeToEthToken'].includes(this.props.swapName) &&
-            <UtxoToAbTexts step={flow.step} flow={flow} swap={swap} coinName={currencyName} />
+            <UtxoToAbTexts flow={flow} swap={swap} stepName={this.getStepName()} />
         }
         {
           ['EthToBtcLike', 'EthTokenToBtcLike'].includes(this.props.swapName) &&
-            <AbToUtxoTexts step={flow.step} flow={flow} swap={swap} coinName={currencyName} />
+            <AbToUtxoTexts flow={flow} swap={swap} stepName={this.getStepName()} />
         }
       </Fragment>
     )
@@ -147,16 +115,16 @@ export default class SwapList extends Component<any, any> {
 
     return (
       <div styleName={`${isMobile ? 'stepList isMobile' : 'stepList'} ${isDark ? 'dark' : ''}`}>
-        {!isMobile && <FirstStep step={flow.step} first={first} second={second} fields={this._fields} text={swapTexts} />}
-        <SecondStep step={flow.step} swap={swap} second={second} windowWidth={windowWidth} fifth={fifth} fourth={fourth} sixth={sixth} fields={this._fields} text={swapTexts} enoughBalance={enoughBalance} />
+        {!isMobile && <FirstStep stepName={this.getStepName()} text={swapTexts} />}
+        {/* <SecondStep step={flow.step} swap={swap} second={second} windowWidth={windowWidth} fifth={fifth} fourth={fourth} sixth={sixth} fields={this._fields} text={swapTexts} enoughBalance={enoughBalance} />
         {
           showDepositWindow &&
             <div styleName="swapDepositWindow">
               <DepositWindow currencyData={currencyData} swap={swap} flow={flow} tokenItems={tokenItems} fields={this._fields} />
             </div>
         }
-        <ThirdStep step={flow.step} windowWidth={windowWidth} swap={swap} sixth={sixth} seventh={seventh} eighth={eighth} fields={this._fields} text={swapTexts} />
-        {!isMobile && <FourthStep step={flow.step} swap={swap} seventh={seventh} eighth={eighth} fields={this._fields} text={swapTexts} />}
+        <ThirdStep step={flow.step} windowWidth={windowWidth} swap={swap} sixth={sixth} seventh={seventh} eighth={eighth} fields={this._fields} text={swapTexts} /> */}
+        {!isMobile && <FourthStep stepName={this.getStepName()} text={swapTexts} />}
       </div>
     )
   }
