+++ src/front/shared/pages/Swap/UTXOSwap/SwapList/SwapSteps/FourthStep.tsx
@@ -9,30 +9,27 @@ import { FormattedMessage } from 'react-intl'
 
 const FourthStep = (props) => {
   const {
-    step,
-    swap,
-    seventh,
-    eighth,
-    fields: {
-      currencyName,
-    },
+    stepName,
     text,
   } = props
 
-  const currencyStep = swap.sellCurrency === currencyName ? seventh : eighth
-  const fourthStepPadding = (step >= currencyStep && isMobile) ? 150 : 0
+  const FOURTH_STEP = ['finish', 'end']
+
+  const isSwapFinished = FOURTH_STEP.includes(stepName)
+
+  const fourthStepPadding = (isSwapFinished && isMobile) ? 150 : 0
 
   return (
     <div
       style={{ paddingTop: fourthStepPadding }}
-      styleName={step >= currencyStep ? 'stepItem active checked' : 'stepItem'}>
-      <span styleName="stepNumber">{step >= currencyStep ? <i className="fas fa-check" /> : 4}</span>
+      styleName={isSwapFinished ? 'stepItem active checked' : 'stepItem'}>
+      <span styleName="stepNumber">{isSwapFinished ? <i className="fas fa-check" /> : 4}</span>
       <p id="swapCompleted" styleName="stepText">
         <FormattedMessage
           id="BtcToEthToken123"
           defaultMessage="The swap is finish!" />
       </p>
-      {step >= currencyStep && (
+      {isSwapFinished && (
         <span styleName="stepHeading">
           {text}
         </span>
